+++ packages/iov-ethereum/src/ethereumconnection.spec.ts
@@ -1950,7 +1950,7 @@ describe("EthereumConnection", () => {
       const querySwapId: AtomicSwapQuery = { id: swapId };
       const querySwapSender: AtomicSwapQuery = { sender: faucetAddress };
       const querySwapRecipient: AtomicSwapQuery = { recipient: recipientAddress };
-      const querySwapHash: AtomicSwapQuery = { hashlock: swapOfferHash };
+      const querySwapHash: AtomicSwapQuery = { hash: swapOfferHash };
 
       // ----- connection.searchTx() -----